move notifications out of misc.
