+++ Fedoraware/TeamFortress2/TeamFortress2.vcxproj
@@ -153,6 +153,7 @@
     <ClCompile Include="TeamFortress2\Features\Camera\CameraWindow.cpp" />
     <ClCompile Include="TeamFortress2\Features\Commands\Commands.cpp" />
     <ClCompile Include="TeamFortress2\Features\CritHack\CritHack.cpp" />
+    <ClCompile Include="TeamFortress2\Features\Misc\Notifications\Notifications.cpp" />
     <ClCompile Include="TeamFortress2\Features\TickHandler\TickHandler.cpp" />
     <ClCompile Include="TeamFortress2\Features\Fedworking\Fedworking.cpp" />
     <ClCompile Include="TeamFortress2\Features\Followbot\Followbot.cpp" />
@@ -329,6 +330,7 @@
     <!-- <ClInclude Include="TeamFortress2\Features\Backtrack\Backtrack.h" /> -->
     <ClInclude Include="TeamFortress2\Features\AutoQueue\AutoQueue.h" />
     <ClInclude Include="TeamFortress2\Features\Commands\Commands.h" />
+    <ClInclude Include="TeamFortress2\Features\Misc\Notifications\Notifications.h" />
     <ClInclude Include="TeamFortress2\Features\TickHandler\TickHandler.h" />
     <ClInclude Include="TeamFortress2\Features\LuaEngine\Callbacks\LuaCallbacks.h" />
     <ClInclude Include="TeamFortress2\Features\LuaEngine\LuaEngine.h" />
